backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
rembg==2.0.50
Pillow==10.2.0
qrcode==7.4.2
requests==2.32.3
onnxruntime==1.18.0
